Create Config Modify module
This is the initial commit of the config modify module. We currently
must overwrite the ConfigInstall class because it doesn't provide a
proper hook to run after optional config is updated.

Similarly we must overwrite the update_helper's `Updater` class because
it doesn't allow executing a CUD without loading it from
`config/update`.

If we can solve those two issues then we can simplify our code quite a
bit.