Improve types for PHPStan
This adds some type annotations and fixes an issue in the
ConfigUpdateDefinition type. We also move the types that we expect to
come from the update_helper module to the `Updater` class and import it
in our `ConfigInstaler` class to re-use it. This allows us to contribute
the types back to that module at a later date.